divide into three parts: handler, repository and service
