+++ posts/repository/user
@@ -1,0 +1,17 @@
+package repository
+
+import (
+	"gorm.io/gorm"
+	"time"
+)
+
+type User struct {
+	Id         int64     `gorm:"column:id"`
+	Name       string    `gorm:"column:name"`
+	CreateTime time.Time `gorm:"column:create_time"`
+	ModifyTime time.Time `gorm:"column:modify_time"`
+}
+
+func (User) TableName() string {
+	return "users"
+}
